requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
discord.py==2.4.0
ez_setup==0.9
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
Jinja2==3.1.4
jsonschema==4.23.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
numpy==2.0.0
packaging==24.1
pillow==10.4.0
pip==22.0.2
protobuf==5.27.2
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
setuptools==70.3.0
six==1.16.0
smmap==5.0.1
streamlit==1.36.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
wheel==0.37.1
yarl==1.9.4
pylint==3.2.5
torch==2.5.0
transformers==4.13.0
sentence-transformers
pinecone-client
yfinance==0.2.54
